@@ -1,6 +1,6 @@
 {
   "name": "react-image-cropper",
-  "version": "1.3.0",
+  "version": "1.3.4",
   "description": "react image crop",
   "main": "app.js",
   "scripts": {
@@ -12,7 +12,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/jerryshew/react-image-cropper"
+    "url": "https://github.com/nearwood/react-image-cropper"
   },
   "keywords": [
     "react",
@@ -23,36 +23,29 @@
   ],
   "author": "jerryshew",
   "license": "MIT",
-  "homepage": "https://github.com",
+  "homepage": "https://github.com/jerryshew",
   "dependencies": {
     "deep-extend": "^0.4.1",
     "npm-run-all": "^4.1.5",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "prop-types": "15.6.0",
+    "react": "16.5.2",
+    "react-dom": "16.5.2"
   },
   "pre-commit": [
     "lint"
   ],
   "devDependencies": {
-    "autoprefixer": "^7.1.1",
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.0.0",
-    "babel-plugin-react-transform": "^1.1.1",
-    "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-runtime": "^6.0.0",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.28.4",
-    "eslint": "^4.16.0",
-    "eslint-config-standard": "^11.0.0-beta.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.6.1",
-    "eslint-plugin-standard": "^3.0.1",
+    "autoprefixer-loader": "^3.1.0",
+    "babel-cli": "^6.10.1",
+    "babel-core": "^5.8.23",
+    "babel-loader": "^5.3.4",
+    "babel-plugin-syntax-dynamic-import": "6.18.0",
+    "babel-plugin-react-transform": "^1.1.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-runtime": "^6.26.0",
+    "cross-env": "^1.0.8",
+    "css-loader": "^0.17.0",
     "express": "^4.13.3",
     "extract-text-webpack-plugin": "^2.1.2",
     "file-loader": "^0.8.5",
