@@ -5,7 +5,7 @@ var config = require('./webpack.config');
 var express = require('express');
 
 var app = new (express)();
-var port = 3000;
+var port = 3676;
 
 var compiler = webpack(config);
 app.use(webpackDevMiddleware(compiler, { noInfo: true, publicPath: config.output.publicPath }));
